+++ FortnoxAPILibrary/Entities/Invoice.cs
@@ -19,12 +19,12 @@ namespace FortnoxAPILibrary.Entities
 
 		/// <remarks/>
 		[JsonProperty]
-		public double? AdministrationFee { get; set; }
+		public decimal? AdministrationFee { get; set; }
 
         /// <summary>This field is Read-Only in Fortnox</summary>
         [ReadOnly]
         [JsonProperty]
-        public double? AdministrationFeeVAT { get; private set; }
+        public decimal? AdministrationFeeVAT { get; private set; }
 
         /// <remarks/>
         [JsonProperty]
@@ -37,12 +37,12 @@ namespace FortnoxAPILibrary.Entities
         /// <summary>This field is Read-Only in Fortnox</summary>
 		[ReadOnly]
 		[JsonProperty]
-		public double? Balance { get; private set; }
+		public decimal? Balance { get; private set; }
 
         /// <summary>This field is Read-Only in Fortnox</summary>
 		[ReadOnly]
 		[JsonProperty]
-		public double? BasisTaxReduction { get; private set; }
+		public decimal? BasisTaxReduction { get; private set; }
 
         /// <summary>This field is Read-Only in Fortnox</summary>
 		[ReadOnly]
@@ -69,12 +69,12 @@ namespace FortnoxAPILibrary.Entities
         /// <summary>This field is Read-Only in Fortnox</summary>
 		[ReadOnly]
 		[JsonProperty]
-		public double? ContributionPercent { get; private set; }
+		public decimal? ContributionPercent { get; private set; }
 
         /// <summary>This field is Read-Only in Fortnox</summary>
 		[ReadOnly]
 		[JsonProperty]
-		public double? ContributionValue { get; private set; }
+		public decimal? ContributionValue { get; private set; }
 
         /// <remarks/>
         [JsonProperty]
@@ -100,11 +100,11 @@ namespace FortnoxAPILibrary.Entities
 
         /// <remarks/>
         [JsonProperty]
-        public double? CurrencyRate { get; set; }
+        public decimal? CurrencyRate { get; set; }
 
         /// <remarks/>
         [JsonProperty]
-        public double? CurrencyUnit { get; set; }
+        public decimal? CurrencyUnit { get; set; }
 
         /// <remarks/>
         [JsonProperty]
@@ -164,17 +164,17 @@ namespace FortnoxAPILibrary.Entities
 
         /// <remarks/>
         [JsonProperty]
-        public double? Freight { get; set; }
+        public decimal? Freight { get; set; }
 
         /// <summary>This field is Read-Only in Fortnox</summary>
         [ReadOnly]
         [JsonProperty]
-        public double? FreightVAT { get; private set; }
+        public decimal? FreightVAT { get; private set; }
 
         /// <summary>This field is Read-Only in Fortnox</summary>
 		[ReadOnly]
 		[JsonProperty]
-		public double? Gross { get; private set; }
+		public decimal? Gross { get; private set; }
 
         /// <summary>This field is Read-Only in Fortnox</summary>
 		[ReadOnly]
@@ -224,7 +224,7 @@ namespace FortnoxAPILibrary.Entities
         /// <summary>This field is Read-Only in Fortnox</summary>
 		[ReadOnly]
 		[JsonProperty]
-		public double? Net { get; private set; }
+		public decimal? Net { get; private set; }
 
         /// <remarks/>
         [JsonProperty]
@@ -289,7 +289,7 @@ namespace FortnoxAPILibrary.Entities
         /// <summary>This field is Read-Only in Fortnox</summary>
 		[ReadOnly]
 		[JsonProperty]
-		public double? RoundOff { get; private set; }
+		public decimal? RoundOff { get; private set; }
 
         /// <summary>This field is Read-Only in Fortnox</summary>
 		[ReadOnly]
@@ -312,7 +312,7 @@ namespace FortnoxAPILibrary.Entities
         /// <summary>This field is Read-Only in Fortnox</summary>
 		[ReadOnly]
 		[JsonProperty]
-		public double? Total { get; private set; }
+		public decimal? Total { get; private set; }
 
         /// <summary>This field is Read-Only in Fortnox</summary>
 		[ReadOnly]
@@ -322,7 +322,7 @@ namespace FortnoxAPILibrary.Entities
         /// <summary>This field is Read-Only in Fortnox</summary>
 		[ReadOnly]
 		[JsonProperty]
-		public double? TotalVAT { get; private set; }
+		public decimal? TotalVAT { get; private set; }
 
         /// <remarks/>
         [JsonProperty]
@@ -458,12 +458,12 @@ namespace FortnoxAPILibrary.Entities
         /// <summary>This field is Read-Only in Fortnox</summary>
 		[ReadOnly]
 		[JsonProperty]
-		public double? ContributionPercent { get; private set; }
+		public decimal? ContributionPercent { get; private set; }
 
         /// <summary>This field is Read-Only in Fortnox</summary>
 		[ReadOnly]
 		[JsonProperty]
-		public double? ContributionValue { get; private set; }
+		public decimal? ContributionValue { get; private set; }
 
         /// <remarks/>
         [JsonProperty]
@@ -471,7 +471,7 @@ namespace FortnoxAPILibrary.Entities
 
         /// <remarks/>
         [JsonProperty]
-        public double? DeliveredQuantity { get; set; }
+        public decimal? DeliveredQuantity { get; set; }
 
         /// <remarks/>
         [JsonProperty]
@@ -479,7 +479,7 @@ namespace FortnoxAPILibrary.Entities
 
         /// <remarks/>
         [JsonProperty]
-        public double? Discount { get; set; }
+        public decimal? Discount { get; set; }
 
         /// <remarks/>
         [JsonProperty]
@@ -499,12 +499,12 @@ namespace FortnoxAPILibrary.Entities
 
 	   /// <remarks/>
 	    [JsonProperty]
-	    public double? Price { get; set; }
+	    public decimal? Price { get; set; }
 
         /// <remarks/>
         [ReadOnly]
         [JsonProperty]
-        public double? PriceExcludingVAT { get; private set; }
+        public decimal? PriceExcludingVAT { get; private set; }
 
         /// <remarks/>
         [JsonProperty]
@@ -513,12 +513,12 @@ namespace FortnoxAPILibrary.Entities
         /// <summary>This field is Read-Only in Fortnox</summary>
 		[ReadOnly]
 		[JsonProperty]
-		public double? Total { get; private set; }
+		public decimal? Total { get; private set; }
 
         /// <remarks/>
         [ReadOnly]
         [JsonProperty]
-        public double? TotalExcludingVAT { get; private set; }
+        public decimal? TotalExcludingVAT { get; private set; }
 
         /// <remarks/>
         [JsonProperty]
@@ -535,7 +535,7 @@ namespace FortnoxAPILibrary.Entities
     {
         /// <remarks/>
         [JsonProperty]
-        public double? Balance { get; set; }
+        public decimal? Balance { get; set; }
 
         /// <remarks/>
         [JsonProperty]
@@ -551,11 +551,11 @@ namespace FortnoxAPILibrary.Entities
 
         /// <remarks/>
         [JsonProperty]
-        public double? CurrencyRate { get; set; }
+        public decimal? CurrencyRate { get; set; }
 
         /// <remarks/>
         [JsonProperty]
-        public double? CurrencyUnit { get; set; }
+        public decimal? CurrencyUnit { get; set; }
 
         /// <remarks/>
         [JsonProperty]
@@ -603,7 +603,7 @@ namespace FortnoxAPILibrary.Entities
 
         /// <remarks/>
         [JsonProperty]
-        public double? Total { get; set; }
+        public decimal? Total { get; set; }
 
         /// <remarks/>
         [JsonProperty]
